server: create a new package dedicated to server configuration
Move the top-level heketi server configuration into a new
package located at server/config.
